lint

+++ src/test/logical_query_plan/lqp_find_subplan_mismatch_test.cpp
@@ -149,8 +149,8 @@ TEST_F(LQPFindSubplanMismatchTest, AdditionalNode) {
 }
 
 TEST_F(LQPFindSubplanMismatchTest, TypeMismatch) {
-  const auto first_node = PredicateNode::make(between(_query_nodes_rhs.table_a_a, (int32_t)42, (int32_t)45));
-  const auto second_node = PredicateNode::make(between(_query_nodes_rhs.table_a_a, (int64_t)42, (int64_t)45));
+  const auto first_node = PredicateNode::make(between(_query_nodes_rhs.table_a_a, 42, 45));
+  const auto second_node = PredicateNode::make(between(_query_nodes_rhs.table_a_a, 42ll, 45ll));
 
   first_node->set_left_input(_query_nodes_rhs.stored_table_node_a);
   second_node->set_left_input(_query_nodes_rhs.stored_table_node_a);
